+++ test/integration/connection_lost_test.exs
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 defmodule ChromicPDF.ConnectionLostTest do
-  use ExUnit.Case, async: false
+  use ChromicPDF.Case, async: false
   alias ChromicPDF.Connection
   alias ChromicPDF.Connection.{ConnectionLostError, Local}