+++ test/integration/zugferd_test.exs
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 defmodule ChromicPDF.ZUGFeRDTest do
-  use ExUnit.Case, async: false
+  use ChromicPDF.Case, async: false
   import ChromicPDF.Utils, only: [system_cmd!: 2]
   require EEx